pyproject.toml
[tool.poetry]
name = "routine-tracker"
version = "0.1.0"
description = "Routine Tracker"
authors = ["Gustavo Barros <gustavosbarros59@gmail.com>"]
license = "README.md"
readme = "README.md"
packages = [{ include = "routine_tracker_core", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
starlette = "0.37.2"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
jwt = "^1.3.1"
responses = "^0.25.3"
fastapi-login = "^1.10.1"
google-auth = "^2.32.0"
python-dotenv = "1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
pytest = "^8.3.2"
pytest-bdd = "^7.2.0"
factory-boy = "^3.3.0"
pytest-freezegun = "^0.4.2"
pytest-mock = "3.11.1"
pytest-cov = "^5.0.0"
pytest-lazy-fixture = "^0.6.3"
pytest-flake8 = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120
skip = ["env", "venv"]

[tool.black]
line-length = 120
